# docs/swagger.yaml
basePath: /
definitions:
  handlers.ServiceHealth:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      service:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  handlers.ServiceStatus:
    properties:
      message:
        type: string
      service:
        type: string
      status:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  models.Coin:
    properties:
      confirmation:
        type: integer
      created_at:
        type: string
      decimal:
        type: integer
      deposit_fee:
        type: string
      deposit_fee_type:
        type: integer
      deposit_gateway:
        items:
          type: string
        type: array
      deposit_status:
        type: integer
      explorer:
        type: string
      explorer_address:
        type: string
      explorer_tx:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      price:
        description: Using string to preserve precision
        type: string
      price_decimal:
        type: integer
      status:
        type: integer
      ticker:
        type: string
      updated_at:
        type: string
      website:
        type: string
      withdraw_fee:
        type: string
      withdraw_fee_type:
        type: integer
      withdraw_gateway:
        items:
          type: string
        type: array
      withdraw_status:
        type: integer
    type: object
  models.CoinListResponse:
    properties:
      coins:
        items:
          $ref: '#/definitions/models.Coin'
        type: array
      total:
        type: integer
    type: object
  models.CoinResponse:
    properties:
      coin:
        $ref: '#/definitions/models.Coin'
    type: object
  models.JWTTokens:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds until expiration
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      message:
        type: string
      redirect_to:
        description: Where to redirect (e.g., "/verify-email")
        type: string
      requires_verify:
        description: True if user needs to verify email
        type: boolean
      tokens:
        $ref: '#/definitions/models.JWTTokens'
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      language:
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      phone_number:
        type: string
      referred_by:
        description: UUID as string in request
        type: string
      timezone:
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  models.RequestOTPRequest:
    properties:
      type:
        enum:
        - email-verification
        - password-reset
        - 2fa
        - phone-verification
        type: string
    required:
    - type
    type: object
  models.RequestOTPResponse:
    properties:
      error:
        description: Only included if email sending failed
        type: string
      expires_in:
        description: Expiration time in seconds
        type: integer
      message:
        type: string
      otp_code:
        description: Only included in development mode when SMTP is disabled
        type: string
    type: object
  models.UserResponse:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_status:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      kyc_status:
        type: string
      language:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      phone_status:
        type: boolean
      role:
        type: string
      status:
        type: string
      timezone:
        type: string
      twofa_enabled:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.VerifyOTPRequest:
    properties:
      code:
        type: string
      type:
        enum:
        - email-verification
        - password-reset
        - 2fa
        - phone-verification
        type: string
    required:
    - code
    - type
    type: object
  models.VerifyOTPResponse:
    properties:
      message:
        type: string
      verified:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: info@bixor.io
  description: High-performance trading backend API for cryptocurrency exchange operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Bixor Trading Engine API
  version: 1.0.0
paths:
  /:
    get:
      consumes:
      - application/json
      description: Get basic information about the Bixor Trading Engine
      produces:
      - application/json
      responses:
        "200":
          description: Landing page message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Landing page
      tags:
      - System
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT tokens
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad request - validation errors
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      summary: User login
      tags:
      - Authorization
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout the current user and invalidate session
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      - BearerAuth: []
      summary: Logout user
      tags:
      - Authorization
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Get current user information based on JWT token
      produces:
      - application/json
      responses:
        "200":
          description: Current user data
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - Authorization
  /api/v1/auth/otp/request:
    post:
      consumes:
      - application/json
      description: Generate and send an OTP code for email verification or other purposes
      parameters:
      - description: OTP request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/models.RequestOTPResponse'
        "400":
          description: Bad request - validation errors or email already verified
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      - BearerAuth: []
      summary: Request an OTP code
      tags:
      - Authorization
  /api/v1/auth/otp/verify:
    post:
      consumes:
      - application/json
      description: Verify an OTP code for email verification or other purposes
      parameters:
      - description: OTP verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            $ref: '#/definitions/models.VerifyOTPResponse'
        "400":
          description: Bad request - validation errors, invalid code, or expired OTP
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: OTP not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      - BearerAuth: []
      summary: Verify an OTP code
      tags:
      - Authorization
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new JWT tokens using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            $ref: '#/definitions/models.JWTTokens'
        "400":
          description: Bad request - validation errors
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      summary: Refresh JWT tokens
      tags:
      - Authorization
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad request - validation errors
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict - user already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BackendSecret: []
      summary: Register a new user
      tags:
      - Authorization
  /api/v1/currency:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all supported cryptocurrencies
      produces:
      - application/json
      responses:
        "200":
          description: List of coins
          schema:
            $ref: '#/definitions/models.CoinListResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get list of all supported coins
      tags:
      - Currency
  /api/v1/currency/{ticker}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific cryptocurrency by
        its ticker symbol
      parameters:
      - description: Coin ticker symbol (e.g., BTC, ETH)
        in: path
        name: ticker
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coin information
          schema:
            $ref: '#/definitions/models.CoinResponse'
        "404":
          description: Coin not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get coin information by ticker
      tags:
      - Currency
  /api/v1/health:
    get:
      consumes:
      - application/json
      description: Check the health status of all services including API and database
      produces:
      - application/json
      responses:
        "200":
          description: All services are healthy
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceHealth'
            type: array
        "503":
          description: One or more services are unhealthy
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceHealth'
            type: array
      summary: Health check for all services
      tags:
      - System
  /api/v1/info:
    get:
      consumes:
      - application/json
      description: Get detailed information about all available services and endpoints
      produces:
      - application/json
      responses:
        "200":
          description: API information with services and endpoints
          schema:
            additionalProperties: true
            type: object
      summary: API information
      tags:
      - System
  /api/v1/status:
    get:
      consumes:
      - application/json
      description: Get the current operational status of all services
      produces:
      - application/json
      responses:
        "200":
          description: All services are active
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceStatus'
            type: array
        "503":
          description: One or more services are inactive
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceStatus'
            type: array
      summary: Get status of all services
      tags:
      - System
schemes:
- http
- https
securityDefinitions:
  BackendSecret:
    description: Backend secret for API authentication (required for protected routes)
    in: header
    name: X-Backend-Secret
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"